indoor_segmentation/requirements.txt
opencv-python==4.6.0.66
pytorch-lightning==1.6.0
easydict
sharedarray==3.2.1
torch_scatter==2.0.9
scipy==1.7.3
matplotlib==3.3.4
tabulate==0.8.9
open3d==0.15.2
einops==0.4.1
timm==0.9.2
plyfile==0.7.4
neptune-client==0.14.2
pytorch-lightning==1.6.0